fix: improve handling of directory buffers
using netrw or nvim-tree to :edit <directory> results in a confusing
buffer with strange properties that had only an icon in the bufferline
because it's name was not being formatted correctly. This change adds an
icon for such buffers and improves their names. Their full mystery
hasn't been resolved yet though...