feat(config)!: add underline indicator style (#514)
This PR refactors the indicator_icon to be a table called indicator, which allows now specifying a style +/- and icon. The only alternative style other than an icon is an underline, which now allows the following appearance. Obviously, a user's mileage will vary depending on their terminal emulator of choice and how it draws underlines.