feat(config): allow specifying style presets
This adds the concept of presets for the bufferline. It's already quite
configurable so I don't intend to have loads but this can be an easy way
to have certain preference e.g. a "minimal" preference which makes the
bg and fg colours the same so only the buffer text is shown

This could also include things like `no_italics` and `no_bold` so there
is an easier way to disable these

chore(config): move and rename highlight resolver

helper function