fix: store paths in g:BufferlinePositions (#780)
This fixes an issue where restoring the custom sort order from a saved session doesn't work correctly most of the time, because it stores buffer ids in g:BufferlinePositions which are not preserved by :mksession. The fix is to store the full paths of the buffers instead. vim.json is required to stringify the list as :mksession only stores string/number values.