docs: fix custom_areas example
Previously a new highlight was created and cached for each index.
For example when the result table only has a warning, an orange color
highlight will be created and cached for index 1. Then the user fixes
the warning but generates an error with index 1, the orange color will be
reused from the cache rather than creating a highlight for the
associated red color. Using an existing highlight with link will
avoid using that cache entirely.